-Werror,-Wdeprecated-anon-enum-enum-conversion
same as f9c6a153c95c4acc6ba02e660c6ca51166b4c79a
"-Werror,-Wdeprecated-anon-enum-enum-conversion"

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/109614
Tested-by: Jenkins

+++ vcl/osx/salmacos.cxx
@@ -345,7 +345,7 @@ bool AquaSalVirtualDevice::SetSize(tools::Long nDX, tools::Long nDY)
         mnBitmapDepth = 32;
         aColorSpace = GetSalData()->mxRGBSpace;
 
-        nFlags = kCGImageAlphaNoneSkipFirst | kCGBitmapByteOrder32Host;
+        nFlags = uint32_t(kCGImageAlphaNoneSkipFirst) | uint32_t(kCGBitmapByteOrder32Host);
     }
 
     // Allocate buffer for virtual device graphics as bitmap context to store graphics with highest required (scaled) resolution
